% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlink-package.R
\docType{package}
\name{gitlink-package}
\alias{gitlink}
\alias{gitlink-package}
\title{gitlink: Add git links to your web-based assets}
\description{
Provides helpers to add git links to Shiny
    applications, R Markdown documents, and other HTML-based resources.
    This is most commonly used for GitHub ribbons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colearendt/gitlink}
  \item Report bugs at \url{https://github.com/colearendt/gitlink/issues}
}

}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
